Fixed the percentage getters

+++ population.cpp
@@ -47,13 +47,13 @@ GenerationData::GenerationData()
     : passiveNumber(0), aggressiveNumber(0), adaptableNumber(0) {}
 
 double GenerationData::GetPassivePercentage() {
-  return (passiveNumber/(passiveNumber + aggressiveNumber + adaptableNumber)) * 100;
+  return (double(passiveNumber)/(passiveNumber + aggressiveNumber + adaptableNumber)) * 100;
 }
 double GenerationData::GetAggressivePercentage() {
   return (double(aggressiveNumber)/(passiveNumber + aggressiveNumber + adaptableNumber)) * 100;
 }
 double GenerationData::GetAdaptablePercentage() {
-  return (adaptableNumber/(passiveNumber + aggressiveNumber + adaptableNumber)) * 100;
+  return (double(adaptableNumber)/(passiveNumber + aggressiveNumber + adaptableNumber)) * 100;
 }
 
 // Singleton Class
